ENH: further refactoring.  Moved common dataframe manipulations out of transforms.py
